Reorganises common directory
