+++ src/common/include/common.h
