+++ src/common/src/column.cpp
@@ -1,4 +1,4 @@
-#include "column.h"
+#include "../include/column.h"
 
 namespace BS {
