+++ src/dirFactory.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "common/dir.h"
+#include "common/include/dir.h"
 
 namespace BS {
     class DirFactory
