+++ src/indiv.h
@@ -7,8 +7,8 @@
 #include <cstdint>
 #include <array>
 // #include "common/basicTypes.h"
-#include "common/coord.h"
-#include "common/dir.h"
+#include "common/include/coord.h"
+#include "common/include/dir.h"
 #include "genome-neurons.h"
 
 namespace BS {
